Fix test cases on release-3.0 (#330)
* tests: allow running a single test

* tests: record the cluster version

* tests: skip 'local' backend tests if cluster is below v4.0.0

* Jenkinsfile: move the CI script into git

* Jenkinsfile: run integration tests in parallel

* tests: allow-auto-random is no longer experimental

* tests: skip local backend better

* Jenkinsfile: moved the file elsewhere

+++ tests/various_types/run.sh
@@ -18,6 +18,9 @@
 set -eu
 
 for BACKEND in importer tidb local; do
+  if [ "$BACKEND" = 'local' ]; then
+    check_cluster_version 4 0 0 'local backend' || continue
+  fi
 
   run_sql 'DROP DATABASE IF EXISTS vt;'
   run_lightning --backend $BACKEND
